# doc/openapi.yaml
openapi: 3.1.0
info:
  title: English Brain API
  version: 0.1.0
  description: |
    API surface for English Brain mobile clients powering onboarding, daily learning sessions,
    personalization, and telemetry. The specification is designed as the single source of truth
    for documentation (Swagger UI/Redoc), SDK generation (openapi-generator), mock servers (Prism),
    test collections (Postman), and contract testing pipelines.
  contact:
    name: English Brain Platform Team
    email: api@englishbrain.app
...
servers:
  - url: https://api.englishbrain.app
    description: Production
  - url: https://staging-api.englishbrain.app
    description: Staging
  - url: http://localhost:8080
    description: Local development
security:
  - FirebaseAuth: []
...
paths:
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        default:
          $ref: '#/components/responses/UnexpectedError'
    patch:
      tags:
        - Users
      summary: Update mutable profile fields
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /users/me/home:
    get:
      tags:
        - Users
      summary: Get home dashboard summary
      operationId: getHomeSummary
      responses:
        '200':
          description: Aggregated home data including goals, streak, and pattern cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeSummary'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /users/me/widget-snapshot:
    get:
      tags:
        - Users
      summary: Get compact snapshot for WidgetKit timelines
      operationId: getWidgetSnapshot
      responses:
        '200':
          description: Widget snapshot data (streak, remaining sentences, next reward)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WidgetSnapshot'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /users/me/tutorial-completions:
    post:
      tags:
        - Onboarding
      summary: Record tutorial completion milestone
      operationId: createTutorialCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TutorialCompletionRequest'
      responses:
        '202':
          description: Tutorial completion accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorialCompletionResponse'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  /users/me/pattern-conquests:
    get:
      tags:
        - Patterns
      summary: Get personalized pattern conquest data
      operationId: getPatternConquests
      responses:
        '200':
          description: Pattern conquest metrics for the user
          content:
            application/json:
              schema:
                type: object
                required:
                  - patterns
                properties:
                  patterns:
                    type: array
                    items:
                      $ref: '#/components/schemas/PatternConquest'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
components:
...
  responses:
    UnexpectedError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Authentication ID token issued for the current user.
  schemas:
    UserProfile:
      type: object
      required:
        - userId
        - email
        - displayName
        - level
        - provisionalLevel
        - timezone
        - locale
        - preferences
        - stats
        - flags
        - createdAt
        - updatedAt
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        level:
          type: integer
          minimum: 1
          maximum: 5
        provisionalLevel:
          type: integer
          minimum: 1
          maximum: 5
          description: Interim level recommendation prior to consolidation with early sessions.
        timezone:
          type: string
          description: IANA timezone identifier.
        locale:
          type: string
          description: BCP 47 language tag.
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        stats:
          $ref: '#/components/schemas/UserStats'
        flags:
          $ref: '#/components/schemas/UserFlags'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserPreferences:
      type: object
      required:
        - hapticsEnabled
        - soundEnabled
        - pushOptIn
        - effectMode
      properties:
        hapticsEnabled:
          type: boolean
        soundEnabled:
          type: boolean
        pushOptIn:
          type: boolean
        effectMode:
          type: string
          enum:
            - full
            - reduced
            - minimal
        dailyGoalSentences:
          type: integer
          minimum: 6
          maximum: 24
        dailyGoalMinutes:
          type: integer
          minimum: 5
          maximum: 30
    UserStats:
      type: object
      required:
        - currentStreak
        - longestStreak
        - brainTokens
        - streakFreezesAvailable
        - patternConquestCount
        - sessionsCompletedThisWeek
        - lastSessionAt
        - subscriptionStatus
      properties:
        currentStreak:
          type: integer
          minimum: 0
        longestStreak:
          type: integer
          minimum: 0
        brainTokens:
          type: integer
          minimum: 0
        streakFreezesAvailable:
          type: integer
          minimum: 0
        patternConquestCount:
          type: integer
          minimum: 0
        sessionsCompletedThisWeek:
          type: integer
          minimum: 0
        lastSessionAt:
          type: string
          format: date-time
          nullable: true
        subscriptionStatus:
          type: string
          enum:
            - free
            - trial
            - premium
    UserFlags:
      type: object
      required:
        - levelTestCompleted
        - tutorialCompleted
        - personalizationReady
      properties:
        levelTestCompleted:
          type: boolean
        tutorialCompleted:
          type: boolean
        personalizationReady:
          type: boolean
        dataDeletionScheduled:
          type: boolean
          default: false
        dataExportRequestedAt:
          type: string
          format: date-time
          nullable: true
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        timezone:
          type: string
        locale:
          type: string
        preferences:
          $ref: '#/components/schemas/UserPreferencesUpdate'
    UserPreferencesUpdate:
      type: object
      properties:
        hapticsEnabled:
          type: boolean
        soundEnabled:
          type: boolean
        pushOptIn:
          type: boolean
        effectMode:
          type: string
          enum:
            - full
            - reduced
            - minimal
        dailyGoalSentences:
          type: integer
          minimum: 6
          maximum: 24
        dailyGoalMinutes:
          type: integer
          minimum: 5
          maximum: 30
    HomeSummary:
      type: object
      required:
        - dailyGoal
        - progress
        - streak
        - brainTokens
        - patternCards
        - recommendedActions
      properties:
        dailyGoal:
          type: object
          required:
            - sentences
            - minutes
            - tier
          properties:
            sentences:
              type: integer
            minutes:
              type: integer
            tier:
              type: string
              enum:
                - basic
                - intensive
              description: Goal pacing tier determined by profile (basic 8–12분 / intensive 12–18분).
        progress:
          type: object
          required:
            - sentencesCompleted
            - minutesSpent
            - lastSessionAt
          properties:
            sentencesCompleted:
              type: integer
            minutesSpent:
              type: integer
            lastSessionAt:
              type: string
              format: date-time
              nullable: true
        streak:
          type: object
          required:
            - current
            - longest
            - freezeEligible
          properties:
            current:
              type: integer
            longest:
              type: integer
            freezeEligible:
              type: boolean
        brainTokens:
          type: object
          required:
            - available
            - pending
          properties:
            available:
              type: integer
            pending:
              type: integer
        patternCards:
          type: array
          items:
            $ref: '#/components/schemas/PatternCard'
        recommendedActions:
          type: array
          items:
            $ref: '#/components/schemas/HomeAction'
        liveActivity:
          $ref: '#/components/schemas/LiveActivityInfo'
    PatternCard:
      type: object
      required:
        - patternId
        - label
        - conquestRate
        - trend
        - severity
        - recommendedAction
      properties:
        patternId:
          type: string
        label:
          type: string
        conquestRate:
          type: number
          minimum: 0
          maximum: 1
        trend:
          type: string
          enum:
            - improving
            - stable
            - declining
        severity:
          type: integer
          minimum: 1
          maximum: 5
        recommendedAction:
          $ref: '#/components/schemas/HomeAction'
        hintRate:
          type: number
          minimum: 0
          maximum: 1
          description: Ratio of attempts using hints within the pattern window.
        firstTryRate:
          type: number
          minimum: 0
          maximum: 1
          description: Share of items solved correctly on the first attempt.
    WidgetSnapshot:
      type: object
      required:
        - updatedAt
        - sentencesRemaining
        - dailyGoalTier
        - currentStreak
        - brainTokens
        - deeplink
      properties:
        updatedAt:
          type: string
          format: date-time
        sentencesRemaining:
          type: integer
          minimum: 0
        dailyGoalTier:
          type: string
          enum:
            - basic
            - intensive
        currentStreak:
          type: integer
          minimum: 0
        brainTokens:
          type: integer
          minimum: 0
        nextBrainTokenInDays:
          type: integer
          minimum: 0
          nullable: true
        deeplink:
          type: string
    HomeAction:
      type: object
      required:
        - type
        - title
        - deeplink
      properties:
        type:
          type: string
          enum:
            - daily-session
            - review
            - brain-burst
            - widget
            - tutorial
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        deeplink:
          type: string
        planRequired:
          type: string
          nullable: true
          enum:
            - free
            - pro
          description: Indicates if the action requires a specific plan tier.
    LiveActivityInfo:
      type: object
      required:
        - supported
        - active
      properties:
        supported:
          type: boolean
          description: Indicates whether the current device is eligible for Live Activity.
        active:
          type: boolean
          description: True when a Live Activity session is currently running.
        activityId:
          type: string
          format: uuid
          nullable: true
        lastUpdatedAt:
          type: string
          format: date-time
          nullable: true
...
    TutorialCompletionRequest:
      type: object
      required:
        - tutorialId
        - completedAt
      properties:
        tutorialId:
          type: string
          example: onboarding-1
        completedAt:
          type: string
          format: date-time
        liveActivityId:
          type: string
          format: uuid
          nullable: true
    TutorialCompletionResponse:
      type: object
      required:
        - tutorialId
        - streakEligible
        - personalizationUnlocked
      properties:
        tutorialId:
          type: string
        streakEligible:
          type: boolean
        personalizationUnlocked:
          type: boolean
...
    PatternConquest:
      type: object
      required:
        - patternId
        - label
        - conquestRate
        - severity
        - exposures
        - lastPracticedAt
        - trend
      properties:
        patternId:
          type: string
        label:
          type: string
        conquestRate:
          type: number
          minimum: 0
          maximum: 1
        severity:
          type: integer
          minimum: 1
          maximum: 5
        exposures:
          type: integer
        lastPracticedAt:
          type: string
          format: date-time
          nullable: true
        trend:
          type: string
          enum:
            - improving
            - stable
            - declining
        hintRate:
          type: number
          minimum: 0
          maximum: 1
        firstTryRate:
          type: number
          minimum: 0
          maximum: 1
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        status:
          type: integer
        details:
          type: array
          items:
            type: object
            additionalProperties: true
...